Create Ticket.cs
Added Ticket class with data models properties to get and set data for ticket and data annotations for front-end display name.
Added new using directives for DataAnnotations
Added Navigation Properties that are "virtual" to allow lazy loading and will return when requested and to signify these items are not stored in the database.