Added Services and Interfaces folder and BTRolesService Interface
Added Services and Interfaces folders to project.
Added BTRolesService Interface withe methods
Added a task method to check to see if use is in role.
Added a task method to return list of user roles.
Added a task method to add user to roles
Added a task method to remove a user from single role.
Added a task method to remove a user from multiple roles.
Added a task method to get users by role and company id
Added a task method to check a user is not in a role by role name and companyId
Added a task method to get roleId from database.