Update BTProjectService.cs
Updated LookupProjectPriorityId to reach out to db and get the priorityId and return it.
Updated GetAllProjectsByPriority to make a list of projects with priorityName from our priorityId with our company Id and return it.
Need to look up issue with why it thinks I'm passing in a string and int and not just an int?