Update BTTicketService.cs
Updated  GetAllTicketsByTypeAsync method to have typeId variable that looks up ticket type name and return it as a value.
Then we try to return a list of tickets where company id's match and select the many tickets and their properties where the ticket type id match the type id to a list and return the list or throw and exception.