Update BTTicketService.cs
Updated GetTicketsByRoleAsync method to create new list of tickets.
Try to see if the user is in role admin by comparing role to roles enum if the user is in admin return the list of ticket by companyid.
Try to see if the user is in role of Developer and get a list of ticket from the db by comparing the company id and the userId to developerId.
Try to see if the user is in role submitter and get a list of tickets by comparing company id where the owner user id and user id match.
See if the user in project manager by comparing role to enum and then get tickets by user id and company id.
Return the list of tickets
If failed catch and thow exception.