Update BTTicketService.cs
Updated BTTicketService with roles Service and project Service.
Updated GetTicketsByUserIdAsync method to get BT user from db and make a new list of tickets.
Try to see what role the user is in. and return a list of tickets for that user in that role to a list.
If you can't catch and throw and exception.