Updated GetTicketDeveloperAsync interface and service.
Updated GetTicketDeveloperAsync method parameters in Interface to include companyId.
Updated GetTicketDeveloperAsync method to be async and pass in the BT users and try to get all the ticked that match ticket ID  and check to see if the developer user id is null if not return the developer user on the ticket. If failed then throw and exception.