Update BTTicketService.cs
Updated AssignTicketAsync to be Async and got the ticket from the db that matches the ticket Id.
Then we try to see if the ticket is null if so we throw and exception.
Then we try to assign the tickets developer User Id to that user Id an look up the ticket status id and return the development value to save to the db if that doesn't work we throw an exception.