Update BTTicketService.cs
Updated the GetProjectTicketsByRoleAsync to be async. Got our list of tickets. Then we are going to try to get tickets by role  using the user and company Id where tickets have a project id that equals the project id passed in and add those tickets to a list.
Return that list and if you can't throw and exceptions.