Update BTTicketHistoryService.cs
Updated AddHistoryAsync method to be asynchronous.
Checked if old ticket or new ticket is not equal to null.
We then are going to take old ticket and compare it to the new ticket and see what has been changed between the old and new ticket.
 We are going to create a TicketHistroy object to send the the db with properties we want in the history.
Then try to add TicketHistories  history to db and save changes,
If it fails throw and exception.
Else if the old ticket is not equal to the new ticket then add new history item based on the difference between the two ticket items.
Then add history to the data set not the db yet.
Then see if the ticket description has changed and add that to the data set.
Then see if the ticket priority has changed and add that to the data set history.
Then see if the ticket status had changed and then add that to the data set history.
Then check ticket type and add changed to data set history.
Then check the developer user id for changes if the oldTicket is null then assign "Not Assigned" to that value and add the history to the data set.
Once we check all our properties for changes between old and new tickets we save the changes to the data base and if that fails we throw and exception.