Update BTTicketHistoryService.cs
Updated the GetCompanyTicketsHistoriesAsyncm method to be asynchronous and run more that one task at a time.
We want to return a list of ticket history.
We are going to create a list of Projects from the companies table that includes projects, tickets, History, user.
We compare the company id to the company id we pass in and send those projects to a list.
Then we are going to get a list of tickets from the many projects and add those tickets to a list.
Then we are going to get a list of ticket history from tickets to add to a list.
Then return the list of ticket histories or if failed throw and exception.