Update TicketHistroy Service Class and Startup.cs
Updated the GetProjectTicketsHistoriesAsync method to be asynchronous.
Try to acess project table where our companies id matches the one we pass in.
Include tickets, ticket histories, user, then get the first or default were project id matches the Id we pass in.
Return a list of ticket history to a list for each project we pass in.
Return the ticket history or if failed catch and throw an exception.

Registered the ticket history interface and implementing class for the service in the startup.cs in the configure services method.