Create BTNotification.cs
Added New Class BTNotification to services using the IBTNotificationService.
Implemented methods from the IBTNotificationService in the BTNotifcation service.
Added injections need for this service to run: ApplicationDBContext, IEmailSender, and IBTRolesService.
Updated AddNotificationAsync to be async and to try to contact the db go to the notification table  and then save changes. If error occurs then throw exception.