Updated IBTNotification and BTNotification service.
Updated SendEmailNotificationAsync method to be async.
Updated Method to be a bool to match the interface dicription.
Set up instance of BTUser to check the user table in the db to see if the user id matches the recipient id we pass in.
Check to see if th btUser is null then return false if it's null.
If we find the btUser then create a string for the btuser email and for the message from notification message.
Then use email service to add btUser email, email subject, and message and return true.
If failed then throw exception.
Else return false.

Updated SendEmailNotificationAsync method discription in interface to be a bool.