Update BTInviteService.cs
Injected our  ApplicationDbContext so we can access out database in this service.
Updated AcceptInviteAsync method to be async and create a invite by checking the Invites db table to see if the Company token and the token we pass in match.
If the Invite is null return false.
Then try if invite is valid make that invite token false so it can't be used again.
Then assign the iviteeId as the userId and save the changes to the db and return true.
If anything fails then throw and exception.