Update BTInviteService.cs
Updated GetInviteAsync method to be async and to get the invite from the Invite db table where the company id matches the one passed in and include the company, project, and invitor and find the first on where the Id matched the invite id passed in.
Return that invite or throw and exception if failed.