Update BTInviteService.cs
Updated GetInviteAsync overloaded method to be async and get invite from db where company id matches the one passed in and include the company, Project, and inviter from the first invite that matches the company token passed in and the invitee email massed in.
Return the invite if failed then throw and exception.