Update BTInviteService.cs
Updated ValidateInviteCodeAsync method to be async and to check if the token in null if so return false.
Set result to false and get invite from Invite data table in db where company token matches the one passed in.
If the invite is not null then we are setting the invite date time and checking to see if the current date time and the invite date is less than 7 days.
Check to see if the invite is valid and set result to true if not return false.