Update BTFileService.cs
Added a private variable to supplement our FormatFileSize method.
Updated ConvertByteArrayToFile to try to covert filedata to a base 64 string and return in this format of the extention of the file then add the file date to the variable imageBase64Data so we can see our file in html.
If failed then throw and exception.