Update BTFileService.cs
Updated ConvertFileToByteArrayAsync to be async.
Try to create an instance of a class MemoryStream from system.IO await the Iformfile we are passing in to the targe stream memoryStream then turn that memoryStream in  byte array named bytefile.
Then close memoryStream and dispose of memory in memoryStream.
The return bytefile if failed then throw and exception.