Update BTFileService.cs
Updated FormatFilesSize to give us the file size with formatted suffix.
We do this by passing in a long data type of bytes, we add an int counter varible set it to zero. We then convert the long  into a decimal.
The we use a while loop and with the math class and the round method to divide the fileSize by 1024 which is 2^10 and while that is greater to or equal to 1 we are going to do something in the loop. Then take the file size divide and keep count of how many times we go through the loop.
Then we return the string and format it show the file size with with one decimal place and the suffix we defined in our private variable.