Updated BTFileService and Startup.cs
Updated GetFileIcon to get icons for fileImage.
set up variable fileImage to default.
See if the string is not null or white space formt he file we pass in.
We then take the variable and set it to the file extension and replacing the . with whitespace.
Then returning the image that match the file in the icon folder.
If not we return the default image.
Commented out registered service in the startup.cs