Updated UserRolesController, BTRolesService, and IBTRolesService
Created UserRolesController constructor and injected services to use in the UserRolesController  to use in my different methods we are going to build.
Added ManageUserRoles method .
Create a new list from ManageUserRolesViewModel then get company id's and get all company users, then loop over useres to add to ViewModel then returnt the model to the view.
Added GetRolesAsync method to BTRolesService to try to get a list of
Added new definiton for method in the IBTRolesService Interface for the GetRolesAsync
Indentityrole from the roles table in the bd and return it if failed throw exception.