Updated the ProjectscController and added a new AddProjectWithPMViewModel
Added new AddProjectWithPMViewModel view model class different property definitions.
Injected the IBTRolesService and IBTLookupService to access our roles and lookup methods we created.
Updated the Create to be a async task and get the companyId from User and add a new AddProjectWithPMViewModel
then load the PMList and the PriorityList and return requested data in the view model.