Updated ProjectsController/Projects Model/ and Projects Create View
Added Service to Porjects controller for IBTLookupService, IBTFileService, and IBTProjectService.
Removed bind from Create method.
Check if model is not null then create instance of companyId from User Identity try to if the ImageFromFile is not null to ConvertFileToByteArrasy get FileName and ContentType then AddNewProjectAsync and add new PM if one was chosen if this fails catch and throw and exception.
Return a redirect to the index page.
Updated DataType in the Project Model for the start and end date.
Added a Null-Conditional Operator for the Project Model in the create view html.