Updated the my ProjectsController/BTProjectService/Archive/Create/Delete/Edit/and Index
Updated ProjectsController Get Edit method
to add companyId variable from the User.Identity.
Added and new model with AddPrjectWithPMViewModel to get project with project Id and Company Id and load a list of PM's and Priority List from the DB.
Added if statement to get the ImageFileData/ ImageFileName/ and ImageFileContentType and Update them in the project.
Changed Delete Method to Archive and updated the methods to use the company Id and Get the projects with User Id and Company Id
Updated the DeleteConfirmed Method to ArchiveConfirmed and again used the User Id, Company Id, and ProjectService to use the ArchiveProjectAsyc method to Archive the project instead of deleteing it.
Added Try Catch block to Archive Project Async Method in BTProejctService.
Fixed the ImageFileData with Razor code in the Porjects View Archive and Changed all Delete name to Archive.
Added Enums Using Directive
Added if statement to check for user is in Role of Adim to change the Project Manager.
Renamed or removed Delete View form the projects Folder.
Added Enums using directive to Projects Edit View and added Method type of post and enctype of multipart form-data.
Hid the Project Id, Company Id, ImageFileData, ImageFileName, ImageFileContentType.
Removed Uneeded Form Groups.
Updated other form-groups with Project to use the ViewModels.AddProjectWithPMViewModel
Added if statement to Preview Image or display a default image.
Updated the Projects Index View to display the Project Image properly and Changed the Delete Button to Archive.