Update TicketsController.cs
Injected new services for the Get Create and Get Post methods.
Added Regions to clean up code and make easier to read.
Refactored Get Create Method to use user manager and add security to check if user is in Admin role or User role and select which data is showed to each.
Refactored the Post Create method to have proper binding that match the ticket views.
Added user manager and security to see if the user is an Admin or user and restrict access to certain data.