Update TicketsController.cs
Added History Service and organized the constructors.
Added old ticket variable to get the Untracked ticket from the DB to compare later to the New ticket after we edit it.
Added new ticket variable with no tracking for comparing to the old ticket.
Added a comparison method for comparing the old ticket and new ticket using the user Id then adding the new history to the DB.